Pull base offscreen rendering into its own class
