+++ yt_idv/rendering_contexts/egl_context.py
@@ -2,12 +2,11 @@ from ctypes import pointer
 
 import numpy as np
 from OpenGL import EGL, GL
-from yt import write_bitmap
 
-from .base_context import BaseContext
+from .base_offscreen import OffscreenRenderingContext
 
 
-class EGLRenderingContext(BaseContext):
+class EGLRenderingContext(OffscreenRenderingContext):
     """Rendering context using EGL (experimental)
 
     Parameters
@@ -77,19 +76,3 @@ class EGLRenderingContext(BaseContext):
 
         GL.glClearColor(0.0, 0.0, 0.0, 0.0)
         GL.glClear(GL.GL_COLOR_BUFFER_BIT | GL.GL_DEPTH_BUFFER_BIT)
-
-    def run(self):
-
-        if self.scene is None:
-            return
-        self.scene.render()
-        if self.image_widget is not None:
-            self.image_widget.value = write_bitmap(self.scene.image[:, :, :3], None)
-            return
-        return self.scene.image
-
-    def snap(self, *args, **kwargs):
-        if self.scene is None:
-            return
-        self.scene.render()
-        super().snap(*args, **kwargs)
